Struct and_around::group::Class

pub struct Class(_);

A group of students. Owns its students.

Methods

impl Class

pub fn new(students: Vec<Rc<Student>>) -> Class

Constructor. Takes a reference counted list of students (should have no other references to them.)

pub fn get_students(&self) -> Vec<Rc<Student>>

Returns a list of references to the students inside.

pub fn get_weak_students(&self) -> Vec<Weak<Student>>

Returns a list of weak references to the students inside. Used primarily for record keeping.

pub fn salt(&mut self)

Shuffles the students in the class.