Struct and_around::group::Group

pub struct Group(_, _);

A temporary group for record keeping, primarily. Weak references so as to allow classes to add and remove students later and still keep past groups with references valid.

Methods

impl Group

pub fn new(size: usize) -> Group

Constructor. Takes a max size.

pub fn try_add(&mut self, student: Rc<Student>) -> Option<Rc<Student>>

Adds a student but checks for capacity first. Returns a None if successful, Some(argument) otherwise.

pub fn contains(&self, student: Rc<Student>) -> bool

Checks if a reference in this grou

pub fn remaining(&self) -> usize

Returns remaining spots for students.

pub fn get_students(&self) -> Vec<Rc<Student>>

Lists students.