Struct and_around::group::GroupManager

pub struct GroupManager { /* fields omitted */ }

Primary struct used for creating, storing, and distributing students among groups.

Methods

impl GroupManager

pub fn new(students: Vec<Rc<Student>>, hist_len: usize) -> GroupManager

Constructor. Takes a list of student references, preferably new. Also takes a length of group states to keep track of.

pub fn distribute(&mut self, sizes: &[usize])

Distribute students without bias. Simply pour them out into the number of groups provided, each with the size provided.

pub fn generate_groups(&mut self, sizes: &[usize])

pub fn push_curr(&mut self)

pub fn show_curr(&self) -> String

Shows the current group state.