Struct and_around::group::History

pub struct History(_);

Contains a circular list of previous groupstates to keep track of student weight.

Methods

impl History

pub fn new(length: usize) -> History

Constructor. Takes a max length of groupstates to keep track of.

pub fn get_weight(&self, s1: Rc<Student>, s2: Rc<Student>) -> u32

Gets a "bias number" based off of the amount of times the students have been grouped together and how long ago it was.

pub fn insert(&mut self, gs: Rc<GroupState>)

Adds a new group state.